projects/parsing/docs-parser/pyproject.toml
[project]
name = "docs-parser"
version = "0.1.0"
description = "Documentation parser for converting verbose docs to AI-dense format"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.12.0",
    "requests>=2.31.0",
    "lxml>=4.9.0",
    "langchain-docling>=0.1.0",
    "langchain-google-community>=1.0.0",
    "tavily-python>=0.3.0",
    "python-dotenv>=1.0.0",
    "protocols",
    "knowledge",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]

[tool.uv.sources]
protocols = { workspace = true }
knowledge = { workspace = true }

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "coverage>=7.10.4",
    "mypy>=1.17.1",
    "pytest>=8.4.1",
    "ruff>=0.12.9",
]

[tool.hatch.build.targets.wheel]
packages = ["src/docs_parser"]
